ISPN-12133 Support XML Fragments in CacheContainerAdmin
